+++ data/meetups.json
@@ -110,5 +110,21 @@
       "andela",
       "music"
     ]
+  },
+  {
+    "id": 8,
+    "createdOn": "2019-01-22 08:41:51",
+    "location": "Buja",
+    "images": [
+      "user.jpg",
+      "test.jpg"
+    ],
+    "topic": "Meetup Title",
+    "description": "Helpers test",
+    "happeningOn": "2019-01-31",
+    "tags": [
+      "andela",
+      "music"
+    ]
   }
 ]
